src/frontend/inpcom.c, Fixing relative path search and NGSPICE_INPUT_DIR with .spiceinit
If we have a project structure like:
  test.cir
  models/unit.mod
  lib/dev.mod

Then if test.cir includes as "models/unit.mod" then the prefix models/ was appended also
in the front of "models/dev.mod" when trying to search for it. So added is search in current
directory, kept the old search, and also appended is the NGSPICE_INPUT_DIR, if valid, to the list
of the sourcepath if it is provided by the .spiceinit, which allow a system wide setup of some
libraries besides the project setup and inclusion of per-project folders.

Another problem was that sourcepath folders were lowered case, which is also fixed.